+++ NEWS.md
@@ -2,6 +2,7 @@ Version 0.12.2-dev
 2012-MM-DD
 
  - Add possibility to specify a literal cache_buster value in benchmark.js
+ - Allow Content-Type with encoding on /layergroup POST (#70)
 
 Version 0.12.0
 2012-04-04